api/requirements.txt
Flask==1.1.2
requests==2.24.0
gunicorn==20.1.0
python-dotenv==0.15.0
Jinja2>=2.9.5
MarkupSafe==0.23
Werkzeug==2.0.3